Added turnout design and control files
These files show how to build a device that will move up to eight servos. It will connect to a JMRI server using C/MRI and the servos can be controlled through JMRI. So far there is no feedback or other features and the turnout open and closed positions will need to be set manually in the Arduino code.